fix compilation for Darwin/MacOSX

+++ Makefile
@@ -48,8 +48,10 @@ else
  endif
 endif
 
-ifeq "$(shell echo 'int main() {return 0; }' | $(CC) -x c - -march=native -o .test 2>/dev/null && echo 1 || echo 0 ; rm -f .test )" "1"
+ifneq "$(shell uname)" "Darwin"
+ ifeq "$(shell echo 'int main() {return 0; }' | $(CC) -x c - -march=native -o .test 2>/dev/null && echo 1 || echo 0 ; rm -f .test )" "1"
 	CFLAGS_OPT = -march=native
+ endif
 endif
 
 ifneq "$(shell uname -m)" "x86_64"
